// BackEnd/BE_jeework/src/app/modules/my-chat/models/message.ts
export  class  Message {
    IdChat: number;
    UserName: string;
    Content_mess: string;
    PathFile: string;
    IsDelAll: boolean;
    IsHidenAll: boolean;
    IsVideoFile: boolean;
    TypeFile: string;
    TenFile: string;
    IdGroup: number;

    //dateRead?: Date;
    CreatedDate: Date;
}